\name{sac-internal}
\title{Internal sac Functions}
\alias{dvalue.epidemic}
\alias{Deltan.ModelTest}
\alias{G.alfa}
\alias{Z.alfa}
\alias{Zn.alfa}
\description{
  Internal speltest functions.
}
\usage{
dvalue.epidemic(x, tol = 1e-07)
Deltan.ModelTest(x, k, m, Alpha, Beta)
Z.alfa(alfa, n, d)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
